feat: parse float and negative number
